feat(QuoteLineRepository): ajout d'une requetes pour calcul du stock actuel pour liste des articles

+++ back/src/main/java/com/mns/cda/saas_facturation/cart/DTO/ActualStockDTO.java
@@ -4,6 +4,6 @@ public record ActualStockDTO(
         Long artId,
         String artName,
         String artRef,
-        int actualStock
+        Long actualStock
 ){
 }
